backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1
PyPDF2==3.0.1
sqlalchemy==2.0.23
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0

